Task_1/requirements.txt
google_api_python_client==2.169.0
python-dotenv==1.1.0
python-telegram-bot==22.0
python-telegram-bot==20.0
google-api-python-client==2.46.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
requests==2.28.1
pandas==1.5.3
